Migrated application to Spring Boot 2
