+++ src/main/java/com/sergtm/controllers/LoginController.java
@@ -1,7 +1,6 @@
 package com.sergtm.controllers;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
