+++ src/main/java/com/sergtm/controllers/ServiceStatusController.java
@@ -1,11 +1,12 @@
 package com.sergtm.controllers;
 
-import com.sergtm.model.Response;
-import com.sergtm.service.IStatusService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.sergtm.model.Response;
+import com.sergtm.service.IStatusService;
 
 @RestController
 @RequestMapping("/status")
