+++ src/main/java/com/sergtm/controllers/WeatherController.java
@@ -1,11 +1,12 @@
 package com.sergtm.controllers;
 
-import com.sergtm.entities.Weather;
-import com.sergtm.service.IWeatherService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.sergtm.entities.Weather;
+import com.sergtm.service.IWeatherService;
 
 @RestController
 @RequestMapping("/weather")
