+++ src/main/java/com/sergtm/dao/impl/StaffMemberDaoImpl.java
@@ -1,17 +1,17 @@
 package com.sergtm.dao.impl;
 
-import com.sergtm.dao.IStaffMemberDao;
-import com.sergtm.entities.StaffMember;
-import com.sergtm.entities.User;
+import java.util.Optional;
+
+import javax.persistence.Query;
+
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.Query;
-
-import java.util.List;
-import java.util.Optional;
+import com.sergtm.dao.IStaffMemberDao;
+import com.sergtm.entities.StaffMember;
+import com.sergtm.entities.User;
 
 @Repository
 @Transactional(readOnly = true)
