+++ src/main/java/com/sergtm/entities/User.java
@@ -1,11 +1,21 @@
 package com.sergtm.entities;
 
-import org.hibernate.annotations.Type;
-
-import javax.persistence.*;
-import java.io.Serializable;
 import java.util.Objects;
 import java.util.Set;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.OneToMany;
+import javax.persistence.SequenceGenerator;
+import javax.persistence.Table;
+
+import org.hibernate.annotations.Type;
 
 @Entity
 @Table( name = "USERS")
