+++ src/main/resources/application.properties
@@ -12,3 +12,7 @@ whetherDataPullerDelay=0 0 * ? * *
 
 #liquibase config
 liquibase.enabled=false
+
+logging.level.org.springframework.web=DEBUG
+#logging.level.com.fasterxml.jackson=DEBUG
+#logging.level.org.hibernate=DEBUG
